requirements.txt
numpy
Network-Script>=1.0.5
Roman-Numerals-Simple
colorama
